wip: open create dialog
