+++ catalog/app/containers/Bucket/Dir.tsx
@@ -7,6 +7,8 @@ import * as RRDom from 'react-router-dom'
 import * as M from '@material-ui/core'
 
 import { Crumb, copyWithoutSpaces, render as renderCrumbs } from 'components/BreadCrumbs'
+import type * as DG from 'components/DataGrid'
+import * as Bookmarks from 'containers/Bookmarks'
 import AsyncResult from 'utils/AsyncResult'
 import * as AWS from 'utils/AWS'
 import * as Config from 'utils/Config'
@@ -15,17 +17,105 @@ import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as BucketPreferences from 'utils/BucketPreferences'
 import parseSearch from 'utils/parseSearch'
-import { getBreadCrumbs, ensureNoSlash, withoutPrefix, up, decode } from 'utils/s3paths'
+import {
+  S3HandleBase,
+  decode,
+  ensureNoSlash,
+  ensureSlash,
+  getBreadCrumbs,
+  up,
+  withoutPrefix,
+} from 'utils/s3paths'
 import type * as workflows from 'utils/workflows'
 
 import Code from './Code'
 import * as FileView from './FileView'
-import { Listing, PrefixFilter } from './Listing'
+import { Item, Listing, PrefixFilter } from './Listing'
 import * as PD from './PackageDialog'
 import * as Successors from './Successors'
 import Summary from './Summary'
 import { displayError } from './errors'
 import * as requests from './requests'
+
+const useHeaderStyles = M.makeStyles((t) => ({
+  root: {
+    alignItems: 'baseline',
+    borderBottom: `1px solid ${t.palette.divider}`,
+    display: 'flex',
+    justifyContent: 'center',
+    padding: t.spacing(0.5, 0),
+  },
+  button: {
+    fontSize: 11,
+    lineHeight: '22px',
+    margin: t.spacing(0, 1),
+  },
+  count: {
+    marginRight: t.spacing(1),
+  },
+  wrapper: {
+    width: '100%',
+  },
+}))
+
+interface HeaderProps {
+  items: Item[]
+  onClearSelection: () => void
+  selection?: DG.GridRowId[]
+}
+
+function Header({ items, onClearSelection, selection }: HeaderProps) {
+  const classes = useHeaderStyles()
+  const count = selection?.length || 0
+  const bookmarks = Bookmarks.use()
+  const bookmarkItems: S3HandleBase[] = React.useMemo(() => {
+    const handles: S3HandleBase[] = []
+    items.some(({ name, handle, type }) => {
+      if (!selection?.length) return true
+      if (selection?.includes(name) && handle)
+        handles.push({
+          ...handle,
+          key: type === 'dir' ? ensureSlash(handle.key) : handle.key,
+        })
+      if (handles.length === selection?.length) return true
+      return false
+    })
+    return handles
+  }, [items, selection])
+  const handleClick = React.useCallback(() => {
+    bookmarkItems?.forEach((handle) => {
+      bookmarks?.append('bookmarks', handle)
+    })
+    onClearSelection()
+  }, [bookmarks, bookmarkItems, onClearSelection])
+  return (
+    <M.Collapse in={!!count} className={classes.wrapper} timeout={100}>
+      <div className={classes.root}>
+        <M.Typography className={classes.count} variant="body2">
+          {count > 1 ? `${count} items are selected` : `${count} item is selected`}
+        </M.Typography>
+        <M.Button
+          className={classes.button}
+          color="primary"
+          size="small"
+          variant="outlined"
+          onClick={handleClick}
+        >
+          Add to bookmarks
+        </M.Button>
+        or
+        <M.Button
+          className={classes.button}
+          color="primary"
+          size="small"
+          onClick={onClearSelection}
+        >
+          Clear selection
+        </M.Button>
+      </div>
+    </M.Collapse>
+  )
+}
 
 interface RouteMap {
   bucketDir: [bucket: string, path?: string, prefix?: string]
@@ -125,11 +215,18 @@ function DirContents({ response, locked, bucket, path, loadMore }: DirContentsPr
         onSelectionChange={handleSelectionModelChange}
         selection={selection}
         toolbarContents={
-          <PrefixFilter
-            key={`${response.bucket}/${response.path}`}
-            prefix={response.prefix}
-            setPrefix={setPrefix}
-          />
+          <>
+            <Header
+              items={items}
+              selection={selection}
+              onClearSelection={() => setSelection([])}
+            />
+            <PrefixFilter
+              key={`${response.bucket}/${response.path}`}
+              prefix={response.prefix}
+              setPrefix={setPrefix}
+            />
+          </>
         }
       />
       {/* Remove TS workaround when Summary will be converted to .tsx */}
