+++ catalog/app/containers/Bucket/Listing.tsx
@@ -9,8 +9,7 @@ import { fade } from '@material-ui/core/styles'
 
 import * as DG from 'components/DataGrid'
 import { renderPageRange } from 'components/Pagination2'
-import * as Bookmarks from 'containers/Bookmarks'
-import * as s3paths from 'utils/s3paths'
+import type { S3HandleBase } from 'utils/s3paths'
 import { readableBytes } from 'utils/string'
 import usePrevious from 'utils/usePrevious'
 
@@ -27,7 +26,7 @@ export interface Item {
   size?: number
   modified?: Date
   archived?: boolean
-  handle?: s3paths.S3HandleBase
+  handle?: S3HandleBase
 }
 
 function maxPartial<T extends R.Ord>(a: T | undefined, b: T | undefined) {
@@ -49,86 +48,6 @@ const computeStats = R.reduce(
     modified: undefined as Date | undefined,
   },
 )
-
-const useHeaderStyles = M.makeStyles((t) => ({
-  root: {
-    alignItems: 'baseline',
-    borderBottom: `1px solid ${t.palette.divider}`,
-    display: 'flex',
-    justifyContent: 'center',
-    padding: t.spacing(0.5, 0),
-  },
-  button: {
-    fontSize: 11,
-    lineHeight: '22px',
-    margin: t.spacing(0, 1),
-  },
-  count: {
-    marginRight: t.spacing(1),
-  },
-  wrapper: {
-    width: '100%',
-  },
-}))
-
-interface HeaderProps {
-  items: Item[]
-  onClearSelection: () => void
-  selection?: DG.GridRowId[]
-}
-
-function Header({ items, onClearSelection, selection }: HeaderProps) {
-  const classes = useHeaderStyles()
-  const count = selection?.length || 0
-  const bookmarks = Bookmarks.use()
-  const bookmarkItems: s3paths.S3HandleBase[] = React.useMemo(() => {
-    const handles: s3paths.S3HandleBase[] = []
-    items.some(({ name, handle, type }) => {
-      if (!selection?.length) return true
-      if (selection?.includes(name) && handle)
-        handles.push({
-          ...handle,
-          key: type === 'dir' ? s3paths.ensureSlash(handle.key) : handle.key,
-        })
-      if (handles.length === selection?.length) return true
-      return false
-    })
-    return handles
-  }, [items, selection])
-  const handleClick = React.useCallback(() => {
-    bookmarkItems?.forEach((handle) => {
-      bookmarks?.append('bookmarks', handle)
-    })
-    onClearSelection()
-  }, [bookmarks, bookmarkItems, onClearSelection])
-  return (
-    <M.Collapse in={!!count} className={classes.wrapper} timeout={100}>
-      <div className={classes.root}>
-        <M.Typography className={classes.count} variant="body2">
-          {count > 1 ? `${count} items are selected` : `${count} item is selected`}
-        </M.Typography>
-        <M.Button
-          className={classes.button}
-          color="primary"
-          size="small"
-          variant="outlined"
-          onClick={handleClick}
-        >
-          Add to bookmarks
-        </M.Button>
-        or
-        <M.Button
-          className={classes.button}
-          color="primary"
-          size="small"
-          onClick={onClearSelection}
-        >
-          Clear selection
-        </M.Button>
-      </div>
-    </M.Collapse>
-  )
-}
 
 interface WrappedAutosizeInputProps extends Omit<AutosizeInputProps, 'ref'> {
   className?: string
@@ -495,8 +414,6 @@ interface ToolbarProps {
   locked?: boolean
   loadMore?: () => void
   items: Item[]
-  selection?: DG.GridRowId[]
-  onClearSelection: () => void
 }
 
 function Toolbar({
@@ -505,13 +422,10 @@ function Toolbar({
   loadMore,
   items,
   children,
-  selection,
-  onClearSelection,
 }: ToolbarProps) {
   const classes = useToolbarStyles()
   return (
     <div className={classes.root}>
-      <Header items={items} selection={selection} onClearSelection={onClearSelection} />
       {children}
       {truncated && (
         <div className={classes.truncated}>
@@ -1122,10 +1036,6 @@ export function Listing({
     [onSelectionChange],
   )
 
-  const onClearSelection = React.useCallback(() => {
-    if (onSelectionChange) onSelectionChange([])
-  }, [onSelectionChange])
-
   // TODO: control page, pageSize, filtering and sorting via props
   return (
     <RootComponent className={cx(classes.root, className)}>
@@ -1143,8 +1053,6 @@ export function Listing({
             loadMore,
             items,
             children: toolbarContents,
-            selection,
-            onClearSelection,
           },
           footer: { truncated, locked, loadMore, items },
         }}
